ATP v3, Validate Compatibility, Improved stop_if, better logs, retry fix (#114)
* Added workflow pre-validation

Also improved debug messages

* Updated go mod/sum

* Progress towards debugging deadlock detection

Mostly this is test cases, and some mild refactoring that should have a similar result

* Use image that supports ARM64

* Ensure that run finishes before closing the step

This ensures that the goroutines are done when the engine exits. This fixes a test case, and will be important when retrying is added to the deadlock check

* Added deadlock check retries

This is a workaround for the situation where all of the steps are processing at the same time, all waiting for input

* Update test cases to take less time

Also improved error logic

* Update dependencies

Required for prior commits to work

* Update go SDK to dev branch

This adds the done message to the SDK

* Updated test deployer to dev branch

* Updated deployers, and added python deployer to run-plugin

* Label loggers

* Update to use ATP v3

This means new error handling, so multiple errors can be reported

* Update SDK in go mod, and add force stop to step interface

Force stop is good for when you do not need to wait for the ATP client, or something equivalant, to finish. Waiting for the ATP to finish means waiting for the steps to finish entirely.

* Cleanup

* Fix tests

* Update SDK and test deployer

* Reduce redundancy in Plugin Provider

* Refactoring to fix linting errors

I reduced the amount of code in one function, and reduced redundancy of unlocking

* Fix linting errors

These ones were only shown on CI, not locally

* bump internal dependencies

* update all internal dependencies

* Remove fragile test

* bump python deployer to 0.3.0

---------

Signed-off-by: Dustin Black <dblack@redhat.com>
Co-authored-by: Dustin Black <dblack@redhat.com>