Deployer types (#117)
* Added workflow pre-validation

Also improved debug messages

* Updated go mod/sum

* Progress towards debugging deadlock detection

Mostly this is test cases, and some mild refactoring that should have a similar result

* Use image that supports ARM64

* Ensure that run finishes before closing the step

This ensures that the goroutines are done when the engine exits. This fixes a test case, and will be important when retrying is added to the deadlock check

* Added deadlock check retries

This is a workaround for the situation where all of the steps are processing at the same time, all waiting for input

* Update test cases to take less time

Also improved error logic

* Update dependencies

Required for prior commits to work

* Update go SDK to dev branch

This adds the done message to the SDK

* Updated test deployer to dev branch

* Updated deployers, and added python deployer to run-plugin

* Label loggers

* Update to use ATP v3

This means new error handling, so multiple errors can be reported

* Update SDK in go mod, and add force stop to step interface

Force stop is good for when you do not need to wait for the ATP client, or something equivalant, to finish. Waiting for the ATP to finish means waiting for the steps to finish entirely.

* Cleanup

* Fix tests

* Update SDK and test deployer

* Reduce redundancy in Plugin Provider

* Refactoring to fix linting errors

I reduced the amount of code in one function, and reduced redundancy of unlocking

* Fix linting errors

These ones were only shown on CI, not locally

* bump internal dependencies

* update all internal dependencies

* refactor config to use multiple local deployers

* config defaults

* fix broken tests

* fix broken tests

* update plugin schema with src and type keys

* remove duplicated test

* add multiple deployer test

* investigate workflow config deployer schema

* refactor deployer discriminator to deployer_id

* refactor deployer discriminator field to use deployer_id

* Progress towards implementation of deployment types

* getting the static type of an object comes out in golang 1.22

* type switch on unserialized  workflow deployer config

* remove dead code

* use pointers

* default to builtin deployer stub

* updated plugin provider schema

* fix some broken tests

* fix more broken tests

* Call close to improve error reporting in Python deployer

* remove dead code

* fix lint and style

* refactor to use deployer_name as discriminator value for factories

* update deployer dependencies

* fix error joining

* unfactor Name back to ID

* unfactor more Name to ID

* update to correct kubernetes deployer version

* fix default value

* Added tests to validate that the workflow fails with invalid deployment configs

* Bump workflow version, and switch from Python to Docker deployer

The docker deployer is already used extensively in the tests. No need to have a second deployer that also needs to be configured

* update readme

---------

Co-authored-by: jaredoconnell <joconnel@redhat.com>
Co-authored-by: Dustin Black <dblack@redhat.com>
Co-authored-by: Matt Leader <mleader@mleader-thinkpadp1gen5.remote.csb>